Minimizing main moosedocs.py script so that it may can be copied to stork
(ref #6699)
